Switch default branch name from `master` to `main`
Github has been using `main` as the default branch name for a while now.
This change updates the branch name in the codebase and the tests so
they work with the default branch name set to `main`.

To make this change locally, you can run the following commands:

    git config --global init.defaultBranch main

I'd suggest also updating the branch name for this repository to `main`
as well and updating the github workflows.